+++ src/VulkanWrapper/Framebuffer.hpp
@@ -35,7 +35,7 @@ public:
         framebufferInfo.sType                   = VK_STRUCTURE_TYPE_FRAMEBUFFER_CREATE_INFO;
         framebufferInfo.flags                   = 0;
         framebufferInfo.renderPass              = renderPass;
-        framebufferInfo.attachmentCount         = attachments.size ();
+        framebufferInfo.attachmentCount         = static_cast<uint32_t> (attachments.size ());
         framebufferInfo.pAttachments            = attachments.data ();
         framebufferInfo.width                   = width;
         framebufferInfo.height                  = height;
